Fix handling of URI errors.
URI errors that were recorded into the repo cache files were not being appropriately handled.  This was happening for RHEL where the key (normally a distribution version such as "8.9") was being coerced to a float in checking if the particular version has support for a specified architecture.  In the case of a URI recorded in the repo cache the contents were not a valid float resulting in an exception.
The fix is to have the base class, Repository, filter out repo entries that indicate an URI error.  A subclass overriding the filter method must first invoke its superclass's filtering after which the subclass may assume any remaining repo entries are of the expected form.

Signed-off-by: Joe Shimkus <jshimkus@redhat.com>